\name{irtGUI}
\alias{irtGUI}
\title{Item Response Theory Analysis with a user-frindly Graphic User Interface.}
\usage{
irtGUI()
}
\value{No return value, launches 'shiny' application}
\description{
Performing Item Response Thory analysis such as paramater estimation, ability estimation, item and model fit analyse, local independence assumption, dimensionality assumption, charachteristic and information curves under various models with a user friendly Graphic User Interface.
}
\examples{
if(interactive()){
## Run this code for launching the Graphic User Interface
irtGUI()
}
}
